impression is now a double

+++ src/models/Word.java
@@ -4,7 +4,7 @@ public class Word {
     public Integer id;
     public String word;
     public Integer count;
-    public Integer impression;
+    public Double impression;
     public Word(String word){
         this.word=word;
     }
